\name{internal}
\alias{pyear}
\alias{pwexp}
\alias{mendelian}
\alias{mendelian.combine}

\title{ Internal functions for marginal method }
\description{Internal functions for marginal method}
\value{
  \item{pyear}{ calculates number of events and person years}
   \item{pwexp}{ estimates survival and hazard for piece-wise exponential model}
   \item{mendelian}{ calculates the mendelian probabilities of carrying the mutation conditional on the proband genotype for 1 gene.}
   \item{mendelian.combine}{ combines mendelian probabilities of carrying the mutation conditional on the proband genotype for 2 genes.}
}

\seealso{ \code{\link{kc.marginal}} }
\keyword{internal}